+++ src/config.ts
@@ -44,7 +44,7 @@ import { LanguageDictionary } from './util/lang';
 import { getQueryParameterByName } from './util/qparam';
 import lang from './resources/lang.json';
 
-const version = '0.4.5-dev.0';
+const version = '0.4.5';
 
 const defaultApiServerUrl = process.env.REACT_APP_XCUBE_API;