+++ CHANGES.md
@@ -1,7 +1,7 @@
 ## Changes in version 0.4.5 (in development)
 
 * Default colour for dataset-related place groups now red while 
-  initial user places colour is colour. (#153)
+  initial user places colour is yellow. (#153)
 * RGB layer no longer hides places. (#152)
 * Users can now download time-series data as a Zip archive
   containing a GeoJSON file for each time-series graph. (#20)