+++ CHANGES.md
@@ -1,4 +1,4 @@
-## Changes in version 0.4.5 (in development)
+## Changes in version 0.4.5
 
 * After logging out, the browser now correctly redirects to the 
   viewer's origin. (#142)